# Micro optimization in Tag Cache